pyproject.toml
[project]
name = "ankilive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.9.1",
    "genanki>=0.13.1",
    "pynput>=1.8.1",
    "markdown>=3.8",
]
